requirement.txt
numpy==1.22.2
onnx==1.15.0
onnxruntime==1.8.0
matplotlib==3.4.1
pillow==8.2.0
torch==1.9.1+cu102
scikit-learn==1.0.2
scikit-opt==0.6.6
pyyaml==5.4.1